Make superrepo work with zuul-cloner
zuul-cloner will assemble something similar to what gerrit's
user-submodule capability assembles, but it won't have any submodules or
be a git repository itself. We still have all of the information we need
without that, but we can't force the checkout of the version we pass in.

This way we have both methods, which will be useful for developers who
will be checking out the result of a gerrit user-submodule in order to
reproduce a build, but CI will have zuul-cloner assembling the
components that will eventually become a commit to the gerrit
user-submodule maintained repository.

+++ giftwrap/tests/utils.py
@@ -35,6 +35,7 @@ def make_test_repo(name='testrepo'):
                 repo.index.commit('test commit')
                 return test(*args, **kwargs)
             finally:
-                shutil.rmtree(testrepo)
+                if os.path.isdir(testrepo):
+                    shutil.rmtree(testrepo)
         return wrapper
     return decorator
